moved DatasetBase to its own module to comply with codefactor (too many lines in module)
